GUI APP - TEXT RECOGNITON/pyproject.toml
[project]
name = "poc-projekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "craft-text-detector>=0.4.1",
    "easyocr>=1.7.2",
    "gitpython>=3.1.44",
    "icecream>=2.1.3",
    "ipython>=8.30.0",
    "mmocr>=1.0.1",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.9.1",
    "paddlepaddle>=2.6.2",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pyqt6-tools>=6.4.2.3.3",
    "qt-material>=2.14",
    "requests>=2.32.3",
    "setuptools>=75.8.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.48.1",
    "ultralytics>=8.3.63",
]
